`connect.php` has your password in it!
Please consider making `connect.php` a private file that is not visibly publicly in the repo. Pascual and Sahar will have to share this file locally.

This password is compromised and needs to be changed ASAP.

Once this is merged, please delete connect.php from the remote repo (but remember to copy it manually to each environment.