fixed 'unlock with password' button and new function
fixed the unlock with password button so that it links to a page that will unlock car with password. Also added a function that checks to see if the password when signing up is at least 10 characters long and only contains letters or numbers. Also created a test page to try to figure out why there is an infinite loop on the signup page.